requirements.txt
python-telegram-bot>=21.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
apscheduler==3.10.4
pytz==2023.3
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9

requests==2.31.0
